Added a DataSource protocol for populating the TEAContributionGraphDelegate.
The data array can be populated by the DataSource connected to the TEAContributionGraph.
Allows users to change the color scheme for the graph as well as the number of grades.
Users can also select which month they want the graph to display which allows more customization.

Updated the demo view to reflect these changes.